Fixed regex and SID
SID was incorrectly documented in the example. SID is MONTHDDYYYYMAPDURATION and not MMDDYYYYMAPDURATION as previously thought. Valorant abbreviates the month.

Forgot to update regex. Now correctly anticipates any number of characters for month abbreviation, any number of numbers for DDYYYY, any number of characters for MAP and exactly 4 numbers for duration. Assumes 10:00 <= match length <=59:99